Proyecto/backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
bcrypt==4.3.0
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1
et_xmlfile==2.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
openpyxl==3.1.5
pillow==11.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
gunicorn